Fix multiple AMD GPU support improperly handling configuration properties

+++ WhateverGreen/kern_igfx.cpp
@@ -481,8 +481,8 @@ void IGFX::wrapCflWriteRegister32(void *that, uint32_t reg, uint32_t value) {
 	} else if (reg == BXT_BLC_PWM_DUTY1) {
 		if (callbackIGFX->driverBacklightFrequency && callbackIGFX->targetBacklightFrequency) {
 			// Translate the PWM duty cycle between the driver scale value and the HW scale value
-			uint32_t rescaledValue = (value * static_cast<uint64_t>(callbackIGFX->targetBacklightFrequency)) /
-				static_cast<uint64_t>(callbackIGFX->driverBacklightFrequency);
+			uint32_t rescaledValue = static_cast<uint32_t>((value * static_cast<uint64_t>(callbackIGFX->targetBacklightFrequency)) /
+				static_cast<uint64_t>(callbackIGFX->driverBacklightFrequency));
 			DBGLOG("igfx", "wrapCflWriteRegister32: write PWM_DUTY1 0x%x/0x%x, rescaled to 0x%x/0x%x", value,
 				   callbackIGFX->driverBacklightFrequency, rescaledValue, callbackIGFX->targetBacklightFrequency);
 			value = rescaledValue;
